+++ python/taco/metadata/sample.py
@@ -1,5 +1,5 @@
 from .derived import GeoEnrich, MajorTOM
-from .spatiotemporal import ISAC, ISTAC, SAC, STAC, TAC
+from .spatiotemporal import ISTAC, STAC, ISpatial, Spatial, Temporal
 from .split import Split
 
-__all__ = ["ISAC", "ISTAC", "SAC", "STAC", "TAC", "GeoEnrich", "MajorTOM", "Split"]
+__all__ = ["ISTAC", "STAC", "GeoEnrich", "ISpatial", "MajorTOM", "Spatial", "Split", "Temporal"]